requirements.txt
PyJWT
redis
boto3
pydantic
python-dotenv
requests
fastapi
mangum
